Flask-restful-showcase/requirements.txt
Flask
Flask-RESTful
flask_testing
flask-cors
mysql-connector-python
python-dotenv
